Add pages controller, define home page route, create partial for logged in user
Created the pages controller which handles the home page which at the "/" route.

Also started to add a partial which will be displayed to logged in users. This has
a "log out" button, the users initials will be displayed in an icon, which will be
clickable to view account details.